refactor: Make the popup a children property of the <VehicleMarker />
