+++ assets/src/components/mapMarkers.tsx
@@ -1,7 +1,7 @@
-import Leaflet, { LatLngExpression, PointTuple } from "leaflet"
+import Leaflet, { LatLngExpression } from "leaflet"
 import "leaflet-defaulticon-compatibility" // see https://github.com/Leaflet/Leaflet/issues/4968#issuecomment-483402699
 import "leaflet.fullscreen"
-import React, { useContext, useRef } from "react"
+import React, { PropsWithChildren, useContext, useRef } from "react"
 import { Marker, Polyline, Popup, Tooltip } from "react-leaflet"
 
 import { StateDispatchContext } from "../contexts/stateDispatchContext"
@@ -28,9 +28,6 @@ import { streetViewUrl } from "../util/streetViewUrl"
 import { TileTypeContext } from "../contexts/tileTypeContext"
 import { ReactMarker } from "./map/utilities/reactMarker"
 import { fullStoryEvent } from "../helpers/fullStory"
-import { DropdownItem, DropdownMenu } from "./map/dropdown"
-import inTestGroup, { TestGroups } from "../userInTestGroup"
-import useScreenSize from "../hooks/useScreenSize"
 
 /*  eslint-enable @typescript-eslint/ban-ts-comment */
 
@@ -99,20 +96,22 @@ const makeLabelIcon = (
   })
 }
 
-export const VehicleMarker = ({
-  vehicle,
-  isPrimary,
-  onSelect,
-  isSelected = false,
-}: {
+interface VehicleMarkerProps extends PropsWithChildren {
   vehicle: Vehicle
   isPrimary: boolean
   isSelected?: boolean
   onSelect?: (vehicle: Vehicle) => void
-}) => {
+}
+
+export const VehicleMarker = ({
+  children,
+  vehicle,
+  isPrimary,
+  onSelect,
+  isSelected = false,
+}: VehicleMarkerProps) => {
   const [{ userSettings }] = useContext(StateDispatchContext)
   const markerRef = useRef<Leaflet.Marker<any>>(null)
-  const screenSize = useScreenSize()
 
   const suppressPopup = () => {
     markerRef.current?.closePopup()
@@ -147,17 +146,6 @@ export const VehicleMarker = ({
   // > [specify] a high value like 1000 [...]
   const zIndexOffset = isSelected ? 1000 : 0
 
-  // This offset is here because, due to a limitation of Leaflet
-  // popups, we weren't able to render the popup at the bottom-right
-  // corner of the marker, where it's supposed to go. This effectively
-  // renders it centered and above the marker, and then uses the
-  // offset to reposition it to the bottom-right corner.
-  const dropdownOffset: PointTuple = [140, 97]
-
-  const dropdownEnabled =
-    inTestGroup(TestGroups.DetoursPilot) &&
-    ["desktop", "tablet"].includes(screenSize)
-
   return (
     <>
       <Marker
@@ -167,15 +155,7 @@ export const VehicleMarker = ({
         zIndexOffset={zIndexOffset}
         ref={markerRef}
       >
-        {dropdownEnabled && (
-          <Popup className="c-dropdown-popup-wrapper" offset={dropdownOffset}>
-            <DropdownMenu>
-              <DropdownItem>
-                Start a detour on route {vehicle.routeId}
-              </DropdownItem>
-            </DropdownMenu>
-          </Popup>
-        )}
+        {children}
       </Marker>
 
       <Marker
