+++ assets/src/components/mapPage/mapDisplay.tsx
@@ -1,4 +1,4 @@
-import Leaflet from "leaflet"
+import Leaflet, { PointTuple } from "leaflet"
 import React, {
   useCallback,
   useContext,
@@ -6,7 +6,7 @@ import React, {
   useState,
   useMemo,
 } from "react"
-import { Pane, useMap } from "react-leaflet"
+import { Pane, Popup, useMap } from "react-leaflet"
 import { SocketContext } from "../../contexts/socketContext"
 import useMostRecentVehicleById from "../../hooks/useMostRecentVehicleById"
 import usePatternsByIdForRoute from "../../hooks/usePatternsByIdForRoute"
@@ -60,6 +60,9 @@ import { LocationType, RouteType } from "../../models/stopData"
 import usePullbackVehicles from "../../hooks/usePullbackVehicles"
 import { fullStoryEvent } from "../../helpers/fullStory"
 import { RecenterControl } from "../map/controls/recenterControl"
+import { DropdownItem, DropdownMenu } from "../map/dropdown"
+import useScreenSize from "../../hooks/useScreenSize"
+import inTestGroup, { TestGroups } from "../../userInTestGroup"
 
 const SecondaryRouteVehicles = ({
   selectedVehicleRoute,
@@ -280,6 +283,19 @@ const SelectedVehicleDataLayers = ({
     useCurrentZoom,
     paddingTopLeft || [0, 0]
   )
+  const screenSize = useScreenSize()
+
+  // This offset is here because, due to a limitation of Leaflet
+  // popups, we weren't able to render the popup at the bottom-right
+  // corner of the marker, where it's supposed to go. This effectively
+  // renders it centered and above the marker, and then uses the
+  // offset to reposition it to the bottom-right corner.
+  const dropdownOffset: PointTuple = [140, 97]
+
+  const dropdownEnabled =
+    inTestGroup(TestGroups.DetoursPilot) &&
+    ["desktop", "tablet"].includes(screenSize)
+
   return (
     <>
       {selectedVehicleOrGhost && (
@@ -292,7 +308,20 @@ const SelectedVehicleDataLayers = ({
                 isPrimary={true}
                 isSelected={true}
                 onSelect={selectVehicle}
-              />
+              >
+                {dropdownEnabled && (
+                  <Popup
+                    className="c-dropdown-popup-wrapper"
+                    offset={dropdownOffset}
+                  >
+                    <DropdownMenu>
+                      <DropdownItem>
+                        Start a detour on route {selectedVehicleOrGhost.routeId}
+                      </DropdownItem>
+                    </DropdownMenu>
+                  </Popup>
+                )}
+              </VehicleMarker>
 
               {!selectedVehicleOrGhost.isShuttle && (
                 <SecondaryRouteVehicles
